requirements.txt
sanic
sanic_cors
sanic_openapi
psycopg2-binary